Move play store launcher bugfix from home to dispatch
